Don't change the window title
We don't clear the title and I don't see an easy way of doing that, so
don't set it at all. Not clearing means that once you use VimRenamer,
your vim instance keeps that title forever (unless you know how to fix
it).

The buffer is given a name, and the default titlestring will use that
name in the title, so I don't think we need a window title too.